[frontend] generic component (#14139)
