+++ opencti-platform/opencti-front/src/private/components/arsenal/vulnerabilities/VulnerabilitySoftwaresDataTable.tsx
@@ -1,0 +1,84 @@
+import { addSoftwaresMutationRelationDelete } from '@components/arsenal/vulnerabilities/AddSoftwaresLines';
+import ListItemText from '@mui/material/ListItemText';
+import { defaultRender } from '../../../../components/dataGrid/dataTableUtils';
+import { truncate } from '../../../../utils/String';
+import { DataTableVariant } from '../../../../components/dataGrid/dataTableTypes';
+import { HexagonOutline } from 'mdi-material-ui';
+import IconButton from '@mui/material/IconButton';
+import { LinkOff } from '@mui/icons-material';
+import DataTableWithoutFragment from '../../../../components/dataGrid/DataTableWithoutFragment';
+import React from 'react';
+import useApiMutation from '../../../../utils/hooks/useApiMutation';
+import { deleteNodeFromEdge } from '../../../../utils/store';
+import { VulnerabilitySoftwaresHas_vulnerability$data } from '@components/arsenal/vulnerabilities/__generated__/VulnerabilitySoftwaresHas_vulnerability.graphql';
+
+type SoftwareType = NonNullable<VulnerabilitySoftwaresHas_vulnerability$data['softwaresHas']>['edges']['0']['node'];
+
+interface VulnerabilitySoftwaresDataTableProps {
+  vulnerabilityId: string;
+  data: SoftwareType[];
+  relationshipType: string;
+}
+
+const VulnerabilitySoftwaresDataTable = ({
+  vulnerabilityId,
+  data,
+  relationshipType,
+}: VulnerabilitySoftwaresDataTableProps) => {
+  const [commit] = useApiMutation(addSoftwaresMutationRelationDelete);
+  const removeSoftware = (event: React.MouseEvent, software: SoftwareType) => {
+    event.stopPropagation();
+    event.preventDefault();
+    commit({
+      variables: {
+        fromId: software.id,
+        toId: vulnerabilityId,
+        relationship_type: relationshipType,
+      },
+      updater: (store) => {
+        deleteNodeFromEdge(store, 'softwares', vulnerabilityId, software.id, { relationshipType, first: 10 });
+      },
+    });
+  };
+
+  return (
+    <DataTableWithoutFragment
+      dataColumns={{
+        name: {
+          percentWidth: 100,
+          isSortable: false,
+          render: ({ name, version, x_opencti_description }) => (
+            <ListItemText
+              primary={defaultRender(`${name} ${
+                version && version.length > 0
+                  ? `(${version})`
+                  : ''
+              }`)}
+              secondary={truncate(x_opencti_description, 120)}
+            />
+          ),
+        },
+      }}
+      data={data}
+      globalCount={data.length}
+      storageKey={`vulnerability-${vulnerabilityId}-softwares-${relationshipType}`}
+      variant={DataTableVariant.inline}
+      hideHeaders
+      getComputeLink={({ id }) => {
+        return `/dashboard/observations/observables/${id}`;
+      }}
+      icon={() => (<HexagonOutline color="primary" />)}
+      actions={(software) => (
+        <IconButton
+          aria-label="Remove"
+          onClick={(event) => removeSoftware(event, software)}
+          size="large"
+        >
+          <LinkOff />
+        </IconButton>
+      )}
+    />
+  );
+};
+
+export default VulnerabilitySoftwaresDataTable;
