--- opencti-platform/opencti-front/src/private/components/arsenal/vulnerabilities/VulnerabilitySoftwaresHas.jsx
@@ -1,114 +1,0 @@
-import React from 'react';
-import { filter } from 'ramda';
-import Typography from '@mui/material/Typography';
-import ListItemText from '@mui/material/ListItemText';
-import IconButton from '@mui/material/IconButton';
-import { LinkOff } from '@mui/icons-material';
-import { HexagonOutline } from 'mdi-material-ui';
-import { graphql, useFragment } from 'react-relay';
-import AddSoftwares from './AddSoftwares';
-import { addSoftwaresMutationRelationDelete } from './AddSoftwaresLines';
-import { commitMutation } from '../../../../relay/environment';
-import { useFormatter } from '../../../../components/i18n';
-import { truncate } from '../../../../utils/String';
-import DataTableWithoutFragment from '../../../../components/dataGrid/DataTableWithoutFragment';
-import { defaultRender } from '../../../../components/dataGrid/dataTableUtils';
-import { DataTableVariant } from '../../../../components/dataGrid/dataTableTypes';
-
-const VulnerabilitySoftwaresHasFragment = graphql`
-  fragment VulnerabilitySoftwaresHas_vulnerability on Vulnerability {
-    id
-    name
-    parent_types
-    entity_type
-    softwaresHas: softwares(relationshipType: "has", first: 10) {
-      edges {
-        node {
-          id
-          parent_types
-          x_opencti_description
-          ... on Software {
-            name
-            version
-            vendor
-          }
-        }
-      }
-    }
-  }
-`;
-
-const VulnerabilitySoftwaresHas = ({ vulnerabilityData }) => {
-  const { t_i18n } = useFormatter();
-  const vulnerability = useFragment(VulnerabilitySoftwaresHasFragment, vulnerabilityData);
-
-  const removeSoftware = (event, software) => {
-    event.stopPropagation();
-    event.preventDefault();
-    commitMutation({
-      mutation: addSoftwaresMutationRelationDelete,
-      variables: {
-        fromId: software.id,
-        toId: vulnerability.id,
-        relationship_type: 'has',
-      },
-      updater: (store) => {
-        const node = store.get(vulnerability.id);
-        const softwares = node.getLinkedRecord('softwares', { relationshipType: 'has', first: 10 });
-        const edges = softwares.getLinkedRecords('edges');
-        const newEdges = filter((n) => n.getLinkedRecord('node').getValue('id') !== software.id, edges);
-        softwares.setLinkedRecords(newEdges, 'edges');
-      },
-    });
-  };
-
-  return (
-    <div style={{ height: '100%', marginTop: 20 }}>
-      <Typography variant="h3" gutterBottom={true} style={{ float: 'left' }}>
-        {t_i18n('Affected software')}
-      </Typography>
-      <AddSoftwares
-        vulnerability={vulnerability}
-        vulnerabilitySoftwares={vulnerability.softwaresHas.edges}
-        relationshipType="has"
-      />
-      <div className="clearfix" />
-      <DataTableWithoutFragment
-        dataColumns={{
-          name: {
-            percentWidth: 100,
-            isSortable: false,
-            render: ({ name, version, x_opencti_description }) => (
-              <ListItemText
-                primary={defaultRender(`${name} ${
-                  version && version.length > 0
-                    ? `(${version})`
-                    : ''
-                }`)}
-                secondary={truncate(x_opencti_description, 120)}
-              />
-            ),
-          },
-        }}
-        data={vulnerability.softwaresHas.edges.map((e) => e.node)}
-        variant={DataTableVariant.inline}
-        hideHeaders
-        getComputeLink={({ id }) => {
-          return `/dashboard/observations/observables/${id}`;
-        }}
-        icon={() => (<HexagonOutline color="primary" />)}
-        actions={(software) => (
-          <IconButton
-            aria-label="Remove"
-            onClick={(event) => removeSoftware(event, software)}
-            size="large"
-          >
-            <LinkOff />
-          </IconButton>
-        )}
-      />
-    </div>
-  );
-};
-
-export default VulnerabilitySoftwaresHas;
