+++ opencti-platform/opencti-front/src/private/components/arsenal/vulnerabilities/VulnerabilitySoftwaresHas.tsx
@@ -1,0 +1,63 @@
+import React from 'react';
+import Typography from '@mui/material/Typography';
+import { graphql, useFragment } from 'react-relay';
+import AddSoftwares from './AddSoftwares';
+import { useFormatter } from '../../../../components/i18n';
+import VulnerabilitySoftwaresDataTable from '@components/arsenal/vulnerabilities/VulnerabilitySoftwaresDataTable';
+import { VulnerabilityDetails_vulnerability$data } from '@components/arsenal/vulnerabilities/__generated__/VulnerabilityDetails_vulnerability.graphql';
+import { VulnerabilitySoftwaresHas_vulnerability$key } from '@components/arsenal/vulnerabilities/__generated__/VulnerabilitySoftwaresHas_vulnerability.graphql';
+
+const VulnerabilitySoftwaresHasFragment = graphql`
+  fragment VulnerabilitySoftwaresHas_vulnerability on Vulnerability {
+    id
+    name
+    parent_types
+    entity_type
+    softwaresHas: softwares(relationshipType: "has", first: 10) {
+      edges {
+        node {
+          id
+          parent_types
+          x_opencti_description
+          ... on Software {
+            name
+            version
+            vendor
+          }
+        }
+      }
+    }
+  }
+`;
+
+interface VulnerabilitySoftwaresHasProps {
+  vulnerabilityData: VulnerabilityDetails_vulnerability$data;
+}
+
+const VulnerabilitySoftwaresHas = ({ vulnerabilityData }: VulnerabilitySoftwaresHasProps) => {
+  const { t_i18n } = useFormatter();
+  const vulnerability = useFragment<VulnerabilitySoftwaresHas_vulnerability$key>(VulnerabilitySoftwaresHasFragment, vulnerabilityData);
+
+  return (
+    <div style={{ height: '100%', marginTop: 20 }}>
+      <Typography variant="h3" gutterBottom={true} style={{ float: 'left' }}>
+        {t_i18n('Affected software')}
+      </Typography>
+      <AddSoftwares
+        vulnerability={vulnerability}
+        vulnerabilitySoftwares={vulnerability.softwaresHas?.edges}
+        relationshipType="has"
+      />
+      <div className="clearfix" />
+      {vulnerability.softwaresHas && (
+        <VulnerabilitySoftwaresDataTable
+          data={vulnerability.softwaresHas.edges.map((e) => e.node)}
+          vulnerabilityId={vulnerability.id}
+          relationshipType="has"
+        />
+      )}
+    </div>
+  );
+};
+
+export default VulnerabilitySoftwaresHas;
