+++ opencti-platform/opencti-front/src/private/components/arsenal/vulnerabilities/VulnerabilitySoftwaresRemediates.tsx
@@ -1,0 +1,66 @@
+import React from 'react';
+import Typography from '@mui/material/Typography';
+import { graphql, useFragment } from 'react-relay';
+import AddSoftwares from './AddSoftwares';
+import { useFormatter } from '../../../../components/i18n';
+import VulnerabilitySoftwaresDataTable from '@components/arsenal/vulnerabilities/VulnerabilitySoftwaresDataTable';
+import { VulnerabilityDetails_vulnerability$data } from '@components/arsenal/vulnerabilities/__generated__/VulnerabilityDetails_vulnerability.graphql';
+import { VulnerabilitySoftwaresRemediates_vulnerability$key } from '@components/arsenal/vulnerabilities/__generated__/VulnerabilitySoftwaresRemediates_vulnerability.graphql';
+
+const VulnerabilitySoftwaresRemediatesFragment = graphql`
+  fragment VulnerabilitySoftwaresRemediates_vulnerability on Vulnerability {
+    id
+    name
+    parent_types
+    entity_type
+    softwaresRemediates: softwares(relationshipType: "remediates", first: 10) {
+      edges {
+        node {
+          id
+          parent_types
+          x_opencti_description
+          ... on Software {
+            name
+            version
+            vendor
+          }
+        }
+      }
+    }
+  }
+`;
+
+interface VulnerabilitySoftwaresRemediatesProps {
+  vulnerabilityData: VulnerabilityDetails_vulnerability$data;
+}
+
+const VulnerabilitySoftwaresRemediates = ({
+  vulnerabilityData,
+}: VulnerabilitySoftwaresRemediatesProps) => {
+  const { t_i18n } = useFormatter();
+
+  const vulnerability = useFragment<VulnerabilitySoftwaresRemediates_vulnerability$key>(VulnerabilitySoftwaresRemediatesFragment, vulnerabilityData);
+
+  return (
+    <div style={{ height: '100%', marginTop: 20 }}>
+      <Typography variant="h3" gutterBottom={true} style={{ float: 'left' }}>
+        {t_i18n('Remediated by')}
+      </Typography>
+      <AddSoftwares
+        vulnerability={vulnerability}
+        vulnerabilitySoftwares={vulnerability.softwaresRemediates?.edges}
+        relationshipType="remediates"
+      />
+      <div className="clearfix" />
+      {vulnerability.softwaresRemediates && (
+        <VulnerabilitySoftwaresDataTable
+          vulnerabilityId={vulnerability.id}
+          data={vulnerability.softwaresRemediates.edges.map((e) => e.node)}
+          relationshipType="remediates"
+        />
+      )}
+    </div>
+  );
+};
+
+export default VulnerabilitySoftwaresRemediates;
